Updates for e26 and c14 qualifiers

+++ icarusalg/Utilities/NonRandomCounter.h
@@ -67,8 +67,11 @@ class util::NonRandomCounter: public CLHEP::HepRandomEngine {
 
   double doFlat()
     {
+      // c14: implicit conversion from 'std::numeric_limits<unsigned long>::type'
+      // (aka 'unsigned long') to 'double' changes value from 
+      // 18446744073709551615 to 18446744073709551616
       return static_cast<double>(++count)
-        / std::numeric_limits<unsigned long>::max();
+        / (double)std::numeric_limits<unsigned long>::max();
     }
